move LuceneQueueEvent
