+++ src/main/java/com/celements/search/lucene/observation/event/LuceneQueueIndexEvent.java
@@ -1,7 +1,6 @@
 package com.celements.search.lucene.observation.event;
 
 import com.celements.common.observation.converter.Remote;
-import com.celements.search.lucene.observation.LuceneQueueEvent;
 
 @Remote
 public class LuceneQueueIndexEvent extends LuceneQueueEvent {
